python/requirements.txt
Django==4.0.4
django-cors-headers==3.12.0
psycopg2==2.9.3
django-environ==0.8.1
jsonpickle==2.2.0